WGS84 Ellipsoid parameters and calculations

Access WGS84 ellipsoid parameters and perform ellipsoid-related calculations including auxiliary latitudes, radii of curvature, and meridian distances.

ellipsoid_params()

ellipsoid_circle(lat)

ellipsoid_latitudes(lat)

ellipsoid_latitudes_inv(lat, type)

ellipsoid_curvature(lat)

Arguments

lat

Numeric vector of geographic (geodetic) latitudes in decimal degrees.

type

Character string specifying the type of auxiliary latitude for inverse conversion. One of: "parametric", "geocentric", "rectifying", "authalic", "conformal", "isometric".

Value

ellipsoid_params(): Named list with WGS84 parameters:

a: Equatorial radius (semi-major axis) in meters

f: Flattening

b: Polar radius (semi-minor axis) in meters

e2: First eccentricity squared

ep2: Second eccentricity squared

n: Third flattening

area: Surface area in square meters

volume: Volume in cubic meters

ellipsoid_circle(): Data frame with columns:

lat: Input latitude

radius: Radius of the circle of latitude in meters

quarter_meridian: Distance from equator to pole along a meridian

meridian_distance: Distance from equator to the given latitude

ellipsoid_latitudes(): Data frame with auxiliary latitudes:

lat: Input geographic latitude

parametric: Parametric (reduced) latitude

geocentric: Geocentric latitude

rectifying: Rectifying latitude

authalic: Authalic latitude

conformal: Conformal latitude

isometric: Isometric latitude

ellipsoid_latitudes_inv(): Data frame with:

input: Input auxiliary latitude

geographic: Corresponding geographic latitude

ellipsoid_curvature(): Data frame with radii of curvature:

lat: Input latitude

meridional: Meridional radius of curvature (M)

transverse: Transverse radius of curvature (N)

Details

The WGS84 ellipsoid is the reference surface used by GPS and most modern mapping systems. It is defined by:

Equatorial radius: 6,378,137 m

Flattening: 1/298.257223563

Auxiliary latitudes are different ways of measuring latitude that are useful in various map projections:

Parametric: Used in ellipsoid parameterization

Geocentric: Angle from center of ellipsoid

Rectifying: Preserves distances along meridians

Authalic: Preserves areas

Conformal: Preserves angles/shapes

Isometric: Used in Mercator projection

Examples

# WGS84 parameters
ellipsoid_params()
#> $a
#> [1] 6378137
#> 
#> $f
#> [1] 0.003352811
#> 
#> $b
#> [1] 6356752
#> 
#> $e2
#> [1] 0.00669438
#> 
#> $ep2
#> [1] 0.006739497
#> 
#> $n
#> [1] 0.00167922
#> 
#> $area
#> [1] 5.100656e+14
#> 
#> $volume
#> [1] 1.083207e+21
#> 

# Radius at different latitudes
ellipsoid_circle(c(0, 30, 45, 60, 90))
#>   lat  radius quarter_meridian meridian_distance
#> 1   0 6378137         10001966                 0
#> 2  30 5528257         10001966           3320113
#> 3  45 4517591         10001966           4984944
#> 4  60 3197105         10001966           6654073
#> 5  90       0         10001966          10001966

# Compare auxiliary latitudes
ellipsoid_latitudes(c(0, 30, 45, 60, 90))
#>   lat parametric geocentric rectifying authalic conformal isometric
#> 1   0    0.00000    0.00000    0.00000  0.00000   0.00000   0.00000
#> 2  30   29.91675   29.83364   29.87515 29.88900  29.83368  31.28104
#> 3  45   44.90379   44.80758   44.85568 44.87170  44.80768  50.22747
#> 4  60   59.91661   59.83308   59.87489 59.88879  59.83322  75.12340
#> 5  90   90.00000   90.00000   90.00000 90.00000  90.00000       Inf

# Radii of curvature
ellipsoid_curvature(c(0, 45, 90))
#>   lat meridional transverse
#> 1   0    6335439    6378137
#> 2  45    6367382    6388838
#> 3  90    6399594    6399594